Move event handlers to new package
Move the event handlers to their own dedicated package to make them easier to find.
