+++ src/main/java/com/maideniles/maidensmaterials/event/BookSpawnEvent.java
@@ -1,4 +1,4 @@
-package com.maideniles.maidensmaterials.init.items.gui;
+package com.maideniles.maidensmaterials.event;
 
 import com.maideniles.maidensmaterials.config.ConfigHandler;
 import com.maideniles.maidensmaterials.init.ItemInit;
@@ -10,20 +10,21 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.PlayerEvent.PlayerLoggedInEvent;
 import net.minecraftforge.items.ItemHandlerHelper;
+
+
 @Mod.EventBusSubscriber
 public class BookSpawnEvent {
 
-	public static final String TAG_PLAYER_HAS_BOOK = "maidensmaterials.book"; 
- 
+	public static final String TAG_PLAYER_HAS_BOOK = "maidensmaterials.book";
+
 	@SubscribeEvent
 	public static void onPlayerLoggedIn(PlayerLoggedInEvent event) {
 		NBTTagCompound playerData = event.player.getEntityData();
-	      NBTTagCompound data = 
-	playerData != null ? playerData.getCompoundTag(EntityPlayer.PERSISTED_NBT_TAG) : new NBTTagCompound();
-	      if(!data.getBoolean(TAG_PLAYER_HAS_BOOK) && (ConfigHandler.BOOK_SPAWN)) {
-	        ItemHandlerHelper.giveItemToPlayer(event.player, new ItemStack(ItemInit.guide_book));
-	        data.setBoolean(TAG_PLAYER_HAS_BOOK, true);
-	        playerData.setTag(EntityPlayer.PERSISTED_NBT_TAG, data);
-	      }
+		NBTTagCompound data = playerData != null ? playerData.getCompoundTag(EntityPlayer.PERSISTED_NBT_TAG) : new NBTTagCompound();
+		if (!data.getBoolean(TAG_PLAYER_HAS_BOOK) && (ConfigHandler.BOOK_SPAWN)) {
+			ItemHandlerHelper.giveItemToPlayer(event.player, new ItemStack(ItemInit.guide_book));
+			data.setBoolean(TAG_PLAYER_HAS_BOOK, true);
+			playerData.setTag(EntityPlayer.PERSISTED_NBT_TAG, data);
+		}
 	}
 }
